Use bicubic interpolation in the 2D canvas (apart from large zooms > 400%, and during the transform animation). This severely improves the display of images by reducing aliasing (some aliasing remains for small zoom factors, where a low-pass prefiltering should be used additionally).
For example, open Barbara and look at the awful aliasing effects of nearest-neighbour interpolation prior to
this patch, and see how the new method improves the situation.

Nearest-neighbour is kept at large zoom levels (>400%), so that the user can recognize that it is operating on
pixels.

During the transform animation, use bilinear interpolation which is faster than bicubic and still much better
than nearest-neighbour.