(IcyExceptionHandler) Locate origin of exceptions in plugin more accurately.
This handles plugin bundles and other corner-cases in two steps. First it finds
the JAR the class comes from, and then the right plugin that was declared on the
repository for this JAR.

With this commit, the error produced by the "GenerateABundleBugPlugin" (beta plugin in Icy)
test plugin is successfully associated with the proper plugin.

This modification allows proper reporting to shared owners of the plugin, whereas before
only the original author was identified and he was the only one receiving the reports.